requirements.txt
streamlit==1.34.0
langchain==0.1.13
langchain-community==0.0.29
huggingface-hub==0.23.2
faiss-cpu==1.8.0
PyDictionary==2.0.1
sympy==1.12.0
sentence-transformers==2.5.1
google-generativeai==0.3.2
langchain-google-genai==0.0.9
PyPDF2==3.0.1
requests==2.31.0
pillow==10.3.0
python-dotenv==1.0.1
pypdf
torch>=2.0.0
diffusers>=0.21.0
transformers>=4.30.0
accelerate>=0.20.0
docx2txt>=0.8